Pref: wb: Move wbListItem to cpp.
